backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic-settings
aiosqlite
email-validator
python-multipart